Address warnings from terraform 0.12 (#26)
refs #24

This partially addresses the issue. When some values are unspecified,
Ironic sets default values and returns the result the next time we
reload the resource. In 0.11, I used 'Default' to have terraform avoid
being surprised, but the better solution is to mark them Computed. This
informs terraform that if the resource doesn't define a value, Ironic
will. This removes most of the warnings in 0.12.