Add timeout option to provider to wait for API
If a timeout is specified, the provider will retry connecting to the API
until the timeout is reached. This can be useful when, for example,
another terraform provider is responsible for creating the Ironic
infrastructure. This is a hack, because terraform doesn't truly
understand depends_on relationships between providers.

This approach has limitations: it only works correctly on the first time
the infrastructure is created. If Ironic goes down you cannot rely on a
subsequent run of the other provider to repair Ironic, as Terraform will
expect to be able to talk to the API during the planning stage. This
isn't relevant to the Metal3 use case, as we only ever run apply once.