Wait for drivers to become available
Currently, the code that polls the Ironic API, only checks if / is
responding.  Once / responds, we should check the drivers endpoint to
see if any drivers are available. This indicates that the conductor is
up and responding as well.  Otherwise, we risk terraform trying to begin
its work with conductor not fully up.