Send config drive data directly to Ironic if it supports it (#35)
Prior to microversion 1.56, you needed to send a base64-encoded gzipped
ISO9660 image containing the config drive contents to Ironic. On newer
versions, we can just send the data directly to Ironic and let it build
the config drive.  By supporting this, we avoid the dependency on
`mkisofs` being available on the system where terraform is run from.