WIP add option to pass user_data url
Currently the installer passes a pointer/wrapper config which
contains a URL, this presents a problem when the target nodes
don't have networking up at the (early) stage of running ignition
in the ramdisk.

Instead, we can download the entire config, which it appears we
can pass to ironic since the instance_info is now defined as
longtext (which on mariadb is 4GB), this is a much higher limit
than the majority of other platforms which do require the pointer
config due to user_data size limits.