provider: refactor API waiting code
This refactors the API waiting code to use ContextWithTimeout to enforce
a real timeout period, instead of an estimate based on the number of
attempts. The estimate approach was wildly inaccurate, and considerably
longer than the user's requested timeout.

It also fixes two critical issues. The first, the waitForAPI function
assumed that it was only waiting for Ironic, not inspector, so it tried
to poll inspector for a list of drivers, which is not a call that would
ever succeed. This splits out the code to two functions; waitForAPI and
waitForConductor.

Second, it works around an issue when accessing the Inspector API
endpoint with a trailing /. Inspector always returns a 404 for `/v1/`,
and the waitForAPI call would always timeout.

To workaround this, we remove the trailing slash for now when polling
the API. This is tracked in
https://storyboard.openstack.org/#!/story/2007660.