provider: do not query drivers for ironic inpsector
We previously encountered issues when trying to provision when the
Ironic API was up, but there were no registered drivers yet -- i.e.,
conductor was not fully online.

We added code to waitForAPI that countered the number of drivers
available and waited for a non-zero amount. However, the waitForAPI code
is used by both inspector and ironic API's. The inspector is never
going to respond to requests for drivers! This was an oversight.