OCPBUGS-8237: rework inspect/clean workflows to avoid an inner loop
Currently, both workflow use a nested flow to work around the inability
to distinguish the initial and final state (manageable). In addition
to being confusing, this inner loop is missing a wait(), resulting
in the get node API being called in a tight loop.

This change uses an internal flag to mark if the process has been
started. This way we can use only one outer loop and implement retries
the same way as for deployment.